moved file according to PR feedback.
