+++ .github/CODEOWNERS
@@ -454,7 +454,7 @@ lerna.json @grafana/frontend-ops
 /public/app/angular/ @torkelo
 /public/app/app.ts @grafana/frontend-ops
 /public/app/dev.ts @grafana/frontend-ops
-/public/app/metrics.ts @grafana/plugins-platform-frontend
+/public/app/core/utils/metrics.ts @grafana/plugins-platform-frontend
 /public/app/index.ts @grafana/frontend-ops
 /public/app/AppWrapper.tsx @grafana/frontend-ops
 /public/app/partials/ @grafana/grafana-frontend-platform
